+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - `LOWER_PRICE_LIMIT`
     - `BLOCK_VOLUME`
     - `VENUE_SPECIFIC_VOLUME_1`
+- Added new publisher for Blue Ocean ATS (`OCEA_MEMOIR_OCEA`)
 
 ## 0.69.0 - 2026-01-13